+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - CollectionView: allowed to render Views inside of specified child Element
 - CollectionView: EmptyView is automatically removed on "addChildView()" calls
 - CollectionView: allowed to be instantiated without inheritance
+- CollectionView: call beforeInit/afterInit on ModelView when several children added
 - Documentation: improve example usage of "OsfObservable.retrigger()" method
 - Documentation: added "async" prefix to examples of View method usages
 - Documentation: fixed formatting in various code examples